Added cars to series

+++ server/models/models.go
@@ -211,15 +211,15 @@ type Result struct {
 }
 
 type Season struct {
-	Id              int    `json:"season_id"`
-	SeasonName      string `json:"season_name"`
-	SeasonShortName string `json:"season_short_name"`
-	SeasonYear      int    `json:"season_year"`
-	SeasonQuarter   int    `json:"season_quarter"`
-	SeriesId        int    `json:"series_id"`
-	LicenseGroup    int    `json:"license_group"`
-	DriverChanges   bool   `json:"driver_changes"`
-	CarClasses      []int  `json:"car_classes"`
+	Id              int     `json:"season_id"`
+	SeasonName      string  `json:"season_name"`
+	SeasonShortName string  `json:"season_short_name"`
+	SeasonYear      int     `json:"season_year"`
+	SeasonQuarter   int     `json:"season_quarter"`
+	SeriesId        int     `json:"series_id"`
+	LicenseGroup    int     `json:"license_group"`
+	DriverChanges   bool    `json:"driver_changes"`
+	CarClasses      []int64 `json:"car_classes"`
 }
 
 type Series struct {
